Add support for the windows key as a modifier.
There was nothing to do really, as the current logic already supported
it. But as the WINDOWS key is not considered an "official" modifier,
shortcuts like "win" were not accepted in the InputTrigger syntax.
This commit just catches this syntax and replace if with the right
KeyEvent.VK_ string prior to calling KeyStroke.getKeyStroke( token ).

So basically, it's a hack.

Oh yeah, and on top of this, it probably won't be that useful.
Indeed, at least on Windows 10, whatever you do the Windows menu
gets called when you release the WINDOWS key, and our Java application
looses the focus.

Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>