InputTriggerDescription has an array of triggers,
instead of a single trigger. This is to facilitate serialization /
deserialization of a single action with several different keybindings
(for instance).

Right now this is incomplete. Because a single InputTriggerDescription
with several triggers will generate correctly several InputTrigger
objects.

However if several InputTriggers are offered even with common contexts
and action but several trigger, each will be built into separate
InputTriggerDescriptions instead of a single one.